+++ Control/Joint.hs
@@ -1,0 +1,6 @@
+module Control.Joint (module Exports) where
+
+import Control.Joint.Schemes as Exports
+import Control.Joint.Effects as Exports
+import Control.Joint.Abilities as Exports
+import Control.Joint.Core as Exports