+++ Control/Joint/Effects/Either.hs
@@ -1,5 +1,6 @@
 module Control.Joint.Effects.Either where
 
+import Control.Joint.Operators ((<$$>), (<**>))
 import Control.Joint.Abilities.Interpreted (Interpreted (Primary, run))
 import Control.Joint.Abilities.Transformer (Transformer (Schema, embed, build, unite), (:>) (T))
 import Control.Joint.Abilities.Liftable (Liftable (lift))
@@ -16,11 +17,11 @@ instance Transformer (Either e) where
 	unite = T . UT
 
 instance Functor u => Functor (UT (Either e) u) where
-	fmap f (UT x) = UT $ (fmap . fmap) f x
+	fmap f (UT x) = UT $ f <$$> x
 
 instance Applicative u => Applicative (UT (Either e) u) where
 	pure = UT . pure . pure
-	UT f <*> UT x = UT $ (<*>) <$> f <*> x
+	UT f <*> UT x = UT $ f <**> x
 
 instance (Applicative u, Monad u) => Monad (UT (Either e) u) where
 	UT x >>= f = UT $ x >>= either (pure . Left) (run . f)