+++ CHANGELOG.md
@@ -35,3 +35,4 @@
 * Lower precedence of `:>` operator to allow write like: `t :> u := a`
 * Create supermodule to have only one line of import in projects
 * Rename `Composition` class to `Interpreted`
+* Add `local` operation for `Reader` effect