+++ CHANGELOG.md
@@ -38,3 +38,4 @@
 * Add `local` operation for `Reader` effect
 * Rename `put` operation for `Writer` to `add`
 * Add `nothing` operation for `Maybe` effect
+* Add `current` operation for `State` effect